backend/requirements.txt
# Flask Backend Dependencies
Flask==3.1.2
flask-cors==5.0.0
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3
openpyxl==3.1.5

# LLM Providers (Phase 2)
openai==1.59.9
google-generativeai==0.8.4

# Image Processing
Pillow==11.1.0
